apps/ms-ia-orquestacion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.10.0
python-dotenv==1.0.0
openai==1.58.1
httpx>=0.27.0,<1.0.0
langchain-text-splitters>=0.3.0,<1.0.0
qdrant-client>=1.11.0,<2.0.0
tiktoken>=0.7.0
pypdf>=5.1.0
